refactor(omegaform): extract toFormSchema into meta/redacted
